TopHalf of ViewRecipe
I made a card Image, with two buttons stacked on the image. The two buttons are a functional back button, and a non functional favorite button.
Added icons and times to the prep/cook time.
Added a ternary expressions to change  the text to people/person.
Fixed the Ingredients and Directions though that will be the next area of rework.